refactor: simplify codebase by removing ~400 lines
Remove unnecessary code and complexity to improve maintainability:

**Quick Wins:**
- Remove unimplemented rename commands (project/branch rename.ts)
- Remove unused yaml dependency from package.json
- Remove all CLI aliases (proj, br, snap, ls, rm, pass)
- Deduplicate formatBytes() - consolidate into helpers.ts

**Big Impact Simplifications:**
- Simplify state lock logic - remove complex PID tracking, keep stale lock detection
- Remove container name cache - Docker API is fast enough
- Remove duplicate state lookup methods (getProject/getBranch nameOrID variants)

**Code Quality:**
- StateManager: Keep only specific methods (getProjectByName, getBranchByNamespace)
- DockerManager: Remove premature optimization caching
- formatBytes: Fix 0 bytes formatting to be consistent ("0.00 B")

**Impact:**
- 400+ lines removed
- 1 dependency removed
- No breaking changes to user-facing commands
- Type checking passes
- Build succeeds
