+++ package.json
@@ -74,7 +74,6 @@
     "commander": "^14.0.1",
     "date-fns": "^4.1.0",
     "dockerode": "^4.0.9",
-    "ora": "^9.0.0",
-    "yaml": "^2.8.1"
+    "ora": "^9.0.0"
   }
 }
